Allow categorical raster hover
